Removed unused import. Extracted http to hpttps replacement to string extension. Extracted the hash logic to an use case.
