+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/domain/extensions/_String.kt
@@ -1,0 +1,6 @@
+package com.pinheiro.marvelhqs.domain.extensions
+
+fun String.httpToHttps(): String {
+    return this.replace("http:", "https:")
+
+}
