+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/domain/usecase/GetCharactersUseCase.kt
@@ -6,13 +6,20 @@ import com.pinheiro.marvelhqs.domain.viewobject.ComicViewObject
 import java.util.concurrent.TimeoutException
 
 class GetCharactersUseCase(
-    private val characterNetworkRepository: ICharacterRepository
+    private val characterNetworkRepository: ICharacterRepository,
+    private val getServerHashUseCase: GetServerHashUseCase
 ) {
-
 
     suspend operator fun invoke(limit: String, offset: String): Result<List<ComicViewObject>> {
         try {
-            val comicDataWrapper = characterNetworkRepository.getCharacter(limit, offset)
+            val ts = "1"
+            val hash = getServerHashUseCase(ts)
+//            val publicKey = BuildConfig.PUBLIC_KEY
+//            val privateKey = BuildConfig.PRIVATE_KEY
+//            val codeToHash = ts + privateKey + publicKey
+//            val md = MessageDigest.getInstance("MD5")
+//            val hash = md.digest(codeToHash.toByteArray()).toHexString()
+            val comicDataWrapper = characterNetworkRepository.getCharacter(limit, offset, ts, hash)
             if (comicDataWrapper.isSuccess) {
 
                 val comicDataContainer = comicDataWrapper.getOrNull()?.data
