+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/domain/usecase/GetFavoriteUseCase.kt
@@ -2,11 +2,7 @@ package com.pinheiro.marvelhqs.domain.usecase
 
 import com.pinheiro.marvelhqs.data.repository.db.interfaces.ICharacterDataBaseRepository
 import com.pinheiro.marvelhqs.data.repository.network.model.ComicDTO
-import com.pinheiro.marvelhqs.domain.mapper.comicDTOListTOComicViewObjectList
-import com.pinheiro.marvelhqs.domain.viewobject.ComicViewObject
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
-import java.io.IOException
 
 class GetFavoriteUseCase(
     val repository: ICharacterDataBaseRepository
